requirements.txt
streamlit
langchain
langchain-community
langchain-openai
pypdf
python-docx
python-dotenv
pandas
